requirements.txt
requests
PyQt6
cryptography
numpy
pandas
pytestPyQt6
python-dotenv
